+++ Sources/SwiftUIKit/Views/MenuList/MenuListText.swift
@@ -9,7 +9,8 @@
 import SwiftUI
 
 /**
- This list item can be used to display several lines of text.
+ This view can be used to display several lines of text in a
+ menu list item.
  */
 public struct MenuListText: View {
     
@@ -31,6 +32,7 @@ public struct MenuListText: View {
 }
 
 struct MenuListText_Previews: PreviewProvider {
+    
     static var previews: some View {
         MenuListText("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.")
     }